backend/requirements.txt
# AdMimic Backend Requirements
# Production-ready dependencies for FastAPI application

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI & Analysis
openai==1.3.7

# Image Processing
Pillow==10.1.0
rembg==2.0.50
colorthief==0.2.1
opencv-python==4.8.1.78

# Data & Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Storage & Database
aiofiles==23.2.1

# HTTP & CORS
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
numpy==1.24.3

# Logging & Monitoring
structlog==23.2.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
